+++ src/Contract/NotificationLoggerInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Neox\WrapNotificatorBundle\Contract;
+
+use Neox\WrapNotificatorBundle\Notification\DeliveryStatus;
+
+interface NotificationLoggerInterface
+{
+    public function log(DeliveryStatus $status): void;
+}