+++ src/Notification/Dto/NotificationDtoInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Neox\WrapNotificatorBundle\Notification\Dto;
+
+interface NotificationDtoInterface
+{
+    public function getChannel(): string;
+}